backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
celery>=5.3
redis>=5.0
django-cors-headers>=4.3
Pillow>=10.0
psycopg2-binary>=2.9
dj-database-url>=2.1
django-storages>=1.14
